% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{choose_number_of_splits}
\alias{choose_number_of_splits}
\title{Identify the number of files to generate given a dataset}
\usage{
choose_number_of_splits(data, dbcon)
}
\arguments{
\item{data}{data.frame}

\item{dbcon}{the database connection}
}
\description{
Per Redshift documentation we're aiming for between 1 MB and 1 GB (after compression).
Since we know almost nothing about how much compression we'll get, we'll target 100 MB
raw sizes with the expectation that we'll compress down to something north of 1 MB and
with the fervent hope that we'll upload files in parallel again at some point.
}
