% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{coalesceifnull}
\alias{coalesceifnull}
\alias{\%||\%}
\title{Simple coalesce}
\usage{
coalesceifnull(x, y)
}
\arguments{
\item{x}{left position}

\item{y}{right position}
}
\description{
A simple non-type aware coalesce
}
