% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{compare_schema_d_to_db}
\alias{compare_schema_d_to_db}
\title{Compare data.frame schema (as detected) to db schema}
\usage{
compare_schema_d_to_db(dbcon, data, table)
}
\arguments{
\item{dbcon}{A database connection object}

\item{data}{data.frame}

\item{table}{character table name}
}
\value{
Side effect of printing result
}
\description{
Compare data.frame schema (as detected) to db schema
}
