% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-spectrum.R
\name{create_external_table_code}
\alias{create_external_table_code}
\title{Define the DDL for an External Table}
\usage{
create_external_table_code(dbcon, d, table_name, location,
  partitioned_by = "", ...)
}
\arguments{
\item{dbcon}{a database connection}

\item{d}{data.frame}

\item{table_name}{Result from dbplyr::in_schema specifying the table and the schema}

\item{location}{s3:// style url}

\item{partitioned_by}{character element containing the column name}
}
\value{
Not specified, dbExecute creates table as side effect
}
\description{
Define the DDL for an External Table
}
