% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{ctas}
\alias{ctas}
\title{Create Table as Select}
\usage{
ctas(query, table_name, ..., temp = TRUE, view = FALSE)
}
\arguments{
\item{query}{The query, either character element or the result of a dplyr chain}

\item{table_name}{The name you want the table to have}

\item{...}{Passed forward to table_attributes}

\item{temp}{boolean.  By default ctas creates temp tables, if you want to override this behavior it takes an explicit named operator.}
}
\description{
Create Table as Select
}
