% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{ctas_code}
\alias{ctas_code}
\title{Create Table as Select Code}
\usage{
ctas_code(query, table_name, temp = TRUE, ...)
}
\arguments{
\item{query}{The query, either character element or the result of a dplyr chain}

\item{table_name}{The name you want the table to have}

\item{temp}{boolean.  By default ctas creates temp tables, if you want to override this behavior it takes an explicit named operator.}

\item{...}{Passed forward to table_attributes}
}
\value{
character. SQL commands to execute the specified create table as select statement
}
\description{
Create Table as Select Code
}
