% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{decompose_in_schema}
\alias{decompose_in_schema}
\title{Decorate in_schema for easy decomposition}
\usage{
decompose_in_schema(table)
}
\arguments{
\item{table}{Character table name (assumed to be in the public schema) or dbplyr::in_schema}
}
\value{
dbplyr::in_schema with attributes 'table_name' and 'schema_name'
}
\description{
Decorate in_schema for easy decomposition
}
