% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-spectrum.R
\name{define_external_table_spec}
\alias{define_external_table_spec}
\title{Define the DDL for an External Table}
\usage{
define_external_table_spec(d, table_name, location, partitioned_by = "")
}
\arguments{
\item{d}{data.frame}

\item{table_name}{Result from dbplyr::in_schema specifying the table and the schema}

\item{location}{s3:// style url}

\item{partitioned_by}{}
}
\description{
Define the DDL for an External Table
}
