% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fix_column_order}
\alias{fix_column_order}
\title{Fix the order of columns in d to match the underlying Redshift table}
\usage{
fix_column_order(d, dbcon, table_name, strict = TRUE)
}
\arguments{
\item{d}{data frame}

\item{dbcon}{db connection}

\item{table_name}{character element}

\item{strict}{boolean}
}
\description{
Internal function for redshiftTools
}
