% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{get_column_names}
\alias{get_column_names}
\title{Title}
\usage{
get_column_names(con, table)
}
\arguments{
\item{con}{DBI database connection}

\item{table}{table_name Either character or dbplyr::in_schema}
}
\value{
character vector of column names
}
\description{
Title
}
