% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{identify_rs_types}
\alias{identify_rs_types}
\title{Translate the types of a given data.frame to Redshift types}
\usage{
identify_rs_types(.data)
}
\arguments{
\item{.data}{data.frame, tbl_df}
}
\value{
character
}
\description{
Translate the types of a given data.frame to Redshift types
}
\examples{
\dontrun{redshiftTools:::identify_rs_types(mtcars)}

}

