% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{is_temp_table}
\alias{is_temp_table}
\title{is_temp_table}
\usage{
is_temp_table(con, table_name)
}
\arguments{
\item{con}{Postgres Connection}

\item{table_name}{Table Name}
}
\value{
boolean
}
\description{
is_temp_table
}
