% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapier-internal.R
\name{mini_cube}
\alias{mini_cube}
\title{Create a mini event_cube}
\usage{
mini_cube(db_fun, name, start, end, resulting_table_name, ...,
  additional_processing = . \%>\% {     . })
}
\arguments{
\item{db_fun}{A R6DB Function}

\item{name}{The event name}

\item{start}{Start Period}

\item{end}{End Period}

\item{...}{Options passed to redshiftTools::ctas to create the table}

\item{additional_processing}{Defined as functional chain, e.g. . %>% identity}
}
\description{
Create a mini event_cube
}
