% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{number_of_slices}
\alias{number_of_slices}
\title{Find the number of slices on the cluster}
\usage{
number_of_slices(dbcon)
}
\description{
We use this as a hint for how many pieces we should slice our data into.
We memoise this query for 1 hour because we may end up doing it several times,
and the result can't change quickly because Redshift resizes take quite some time.
}
