% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-errors.R
\name{recent_errors}
\alias{recent_errors}
\title{Find recent redshift errors}
\usage{
recent_errors(con, n = 10)
}
\arguments{
\item{con}{Redshift connection}

\item{n}{Number of errors to show}
}
\value{
data.frame
}
\description{
Newest errors at the top
}
