% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_add_column.R
\name{rs_add_column}
\alias{rs_add_column}
\title{Add a column to a redshift table}
\usage{
rs_add_column(dbcon, table_name, column_name, redshift_type)
}
\arguments{
\item{dbcon}{Database connection object of type RPostgreSQL}

\item{table_name}{the name of the target table (character)}

\item{column_name}{(character)}
}
\description{
Add a typed column to a redshift database
}

