% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert2.R
\name{rs_cols_upsert_table}
\alias{rs_cols_upsert_table}
\title{Upsert Redshift table by specifying a column list}
\usage{
rs_cols_upsert_table(dat, dbcon, table_name, values, keys, split_files,
  bucket = Sys.getenv("AWS_BUCKET_NAME"),
  region = Sys.getenv("AWS_DEFAULT_REGION"),
  access_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  iam_role_arn = Sys.getenv("AWS_IAM_ROLE_ARN"), wlm_slots = 1,
  additional_params = "")
}
\arguments{
\item{dat}{a data frame}

\item{dbcon}{connection to the Redshift server}

\item{table_name}{the name of the table to replace}

\item{values}{the columns that reflect updated values}

\item{keys}{the key columns on which to upsert}

\item{split_files}{how many files to split into. Default based on number of slices in Redshift.}

\item{bucket}{temporary AWS bucket; defaults to AWS_BUCKET_NAME}

\item{region}{temporary AWS region; defaults to AWS_DEFAULT_REGION}

\item{access_key}{AWS access key; defaults to AWS_ACCESS_KEY_ID}

\item{secret_key}{AWS secret key; defaults to AWS_SECRET_ACCESS_KEY}

\item{iam_role_arn}{IAM role; defaults to AWS_IAM_ROLE_ARN. Ignores access_key and secret_key.}

\item{wlm_slots}{number of WLM slots for bulk load}

\item{additional_params}{Additional paramseters to Redshift COPY statement}
}
\description{
Upload a table to S3 and then load to Redshift, replacing the target value
in existing rows that have the same keys, and inserting rows with new keys.
New rows must match structure and column ordering of existing Redshift table.
}
\examples{
library(DBI)

a=data.frame(a=seq(1,10000), b=seq(10000,1))
n=head(a,n=5000)
n$b=n$a
nx=rbind(n, data.frame(a=seq(99999:104000), b=seq(104000:99999)))

\dontrun{
con <- dbConnect(RPostgres::Postgres(), dbname="dbname",
host='my-redshift-url.amazon.com', port='5439',
user='myuser', password='mypassword',sslmode='require')

rs_cols_upsert_table(df=nx, dbcon=con, table_name='testTable',
bucket="my-bucket", split_files=4, values=c('b'), keys=c('a'))
}
}
\seealso{
\url{https://docs.aws.amazon.com/redshift/latest/dg/merge-specify-a-column-list.html}

\url{http://docs.aws.amazon.com/redshift/latest/dg/tutorial-configuring-workload-management.html}
}
