% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_definition.R
\name{rs_create_statement}
\alias{rs_create_statement}
\title{Generate create table statement for Amazon Redshift}
\usage{
rs_create_statement(df, tableName = deparse(substitute(df)))
}
\arguments{
\item{df}{the data.frame you want to upload to Amazon Redshift}

\item{tableName}{the name of the table to create, if not specified it'll use the data.frame name}
}
\description{
This lets you easily generate a table schema from a data.frame, which allows for easily uploading to redshift afterwards.
}
\examples{

n=1000
testdf = data.frame(
a=rep('a', n),
b=c(1:n),
c=rep(as.Date('2017-01-01'), n),
d=rep(as.POSIXct('2017-01-01 20:01:32'), n),
e=rep(as.POSIXlt('2017-01-01 20:01:32'), n),
f=rep(paste0(rep('a', 4000), collapse=''), n) )

cat(getRedshiftCreateStatement(testdf, tablename='dm_great_table'))

}
