% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-create.R
\name{rs_create_table}
\alias{rs_create_table}
\title{Create redshift table}
\usage{
rs_create_table(.data, dbcon, table_name)
}
\arguments{
\item{.data}{\code{data.frame}}

\item{table_name}{\code{character}}
}
\value{
The column names that redshift actually ended up using
}
\description{
Create redshift table
}

