% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-create.R
\name{rs_create_table_spec}
\alias{rs_create_table_spec}
\title{Creates an Redshift column spec given a data.frame}
\usage{
rs_create_table_spec(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
character element wrapped in () reflecting the infered table spec
}
\description{
Creates an Redshift column spec given a data.frame
}
