% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_utils.R
\name{rs_list_tables}
\alias{rs_list_tables}
\title{List redshift tables}
\usage{
rs_list_tables(dbcon, pattern = ".*")
}
\arguments{
\item{dbcon}{an RPostgres/RJDBC connection to the redshift server}

\item{pattern}{regular expression to match subset. defaults to all tables}
}
\value{
a character vector of table names in the format {schema name}.{table name}
}
\description{
Lists all tables available in redshift database
}
