% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{rs_move_table}
\alias{rs_move_table}
\title{Move redshift table from one cluster to another}
\usage{
rs_move_table(table_name, schema_from, schema_to, dbcon_from, dbcon_to,
  split_files, bucket = Sys.getenv("AWS_BUCKET_NAME"),
  region = Sys.getenv("AWS_DEFAULT_REGION"),
  access_key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  session_token = Sys.getenv("AWS_SESSION_TOKEN"),
  iam_role_arn = Sys.getenv("AWS_IAM_ROLE_ARN"), wlm_slots = 1,
  additional_params = "")
}
\arguments{
\item{table_name}{the name of the table to move}

\item{schema_from}{the name of the schema where the source table is}

\item{schema_to}{the name of the destination schema}

\item{dbcon_from}{a db connection to the source RS cluster}

\item{dbcon_to}{a db connection to the destination RS cluster}

\item{split_files}{optional parameter to specify amount of files to split into. If not specified will look at amount of slices in Redshift to determine an optimal amount.}

\item{bucket}{the name of the temporary bucket to load the data. Will look for AWS_BUCKET_NAME on environment if not specified.}

\item{region}{the region of the bucket. Will look for AWS_DEFAULT_REGION on environment if not specified.}

\item{access_key}{the access key with permissions for the bucket. Will look for AWS_ACCESS_KEY_ID on environment if not specified.}

\item{secret_key}{the secret key with permissions fot the bucket. Will look for AWS_SECRET_ACCESS_KEY on environment if not specified.}

\item{session_token}{the session key with permissions for the bucket, this will be used instead of the access/secret keys if specified. Will look for AWS_SESSION_TOKEN on environment if not specified.}

\item{iam_role_arn}{an iam role arn with permissions fot the bucket. Will look for AWS_IAM_ROLE_ARN on environment if not specified. This is ignoring access_key and secret_key if set.}

\item{wlm_slots}{amount of WLM slots to use for this bulk load http://docs.aws.amazon.com/redshift/latest/dg/tutorial-configuring-workload-management.html}

\item{additional_params}{Additional params to send to the COPY statement in Redshift}
}
\description{
Unload a table to S3 and then load it with redshift, replacing the contents of that table.
The table on redshift has to have the same structure and column ordering to work correctly.
}
\examples{
library(DBI)

a=data.frame(a=seq(1,10000), b=seq(10000,1))

\dontrun{
dbcon_from <- dbConnect(RPostgres::Postgres(), dbname="dbname",
host='my-redshift-url.amazon.com', port='5439',
user='myuser', password='mypassword',sslmode='require')
 dbcon_to <- dbConnect(RPostgres::Postgres(), dbname="dbname2",
host='my-redshift-url2.amazon.com', port='5439',
user='myuser', password='mypassword2',sslmode='require')

rs_move_table(table_name,"public","public",dbcon_from,dbcom_to,
bucket="my-bucket", split_files=4)

}
}
