% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{rs_table_exists}
\alias{rs_table_exists}
\title{Check if table exists}
\usage{
rs_table_exists(dbcon, table_name)
}
\arguments{
\item{dbcon}{database connection}

\item{table_name}{dbplyr::in_schema result}
}
\value{
boolean
}
\description{
Passing this check does not mean that the user has access to the table per se.
}
