% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{rs_unload}
\alias{rs_unload}
\title{Unload data from Redshift}
\usage{
rs_unload(con, query, s3_prefix,
  options = "DELIMITER ',' BZIP2 ADDQUOTES ESCAPE PARALLEL ON")
}
\arguments{
\item{con}{DBI connection object}

\item{query}{character query}

\item{s3_prefix}{character; s3 url with trailing slash prefix}

\item{options}{As in https://docs.aws.amazon.com/redshift/latest/dg/r_UNLOAD.html}
}
\value{
undefined
}
\description{
Unload data from Redshift
}
