% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{sanitize_column_names_for_redshift}
\alias{sanitize_column_names_for_redshift}
\title{Make Column Names that Redshift Can Be Happy With}
\usage{
sanitize_column_names_for_redshift(.data, rw_prefix = TRUE)
}
\arguments{
\item{.data}{data.frame or character vector}
}
\value{
character vector
}
\description{
Make Column Names that Redshift Can Be Happy With
}
