% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-spectrum.R
\name{spectrum_add_partition}
\alias{spectrum_add_partition}
\title{Add a partition to Redshift Spectrum}
\usage{
spectrum_add_partition(dbcon, table_name, part_name, part_value, base_location,
  quote_partition_value = TRUE)
}
\arguments{
\item{dbcon}{The database connection to Redshift}

\item{table_name}{A dblyr::in_schema defined table name}

\item{part_name}{The name of the partition}

\item{part_value}{The value of the partition}

\item{base_location}{The base s3:// location of the partition files. It is assumed that the partition is specified as {base_location}/{part_name}={part_value}}
}
\description{
Add a partition to Redshift Spectrum
}
