% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-spectrum.R
\name{spectrum_drop_partition}
\alias{spectrum_drop_partition}
\title{Remove a partition from Redshift Spectrum}
\usage{
spectrum_drop_partition(dbcon, table_name, part_name, part_value)
}
\arguments{
\item{dbcon}{The database connection to Redshift}

\item{table_name}{A dblyr::in_schema defined table name}

\item{part_name}{The name of the partition}

\item{part_value}{The value of the partition}
}
\description{
Remove a partition from Redshift Spectrum
}
