% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-spectrum.R
\name{spectrum_list_partitions}
\alias{spectrum_list_partitions}
\title{List Spectrum partitions}
\usage{
spectrum_list_partitions(dbcon, table_name, stop_if_empty = TRUE)
}
\arguments{
\item{dbcon}{regular db connection object}

\item{table_name}{dbplyr::in_schema specification of table name}
}
\description{
Provides character vector of partition names for a given table
}
