% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{table_attributes}
\alias{table_attributes}
\title{Get Table Attributes String}
\usage{
table_attributes(diststyle = c("even", "all", "key"), distkey = NULL,
  compound_sort = NULL, interleaved_sort = NULL)
}
\arguments{
\item{diststyle}{Distribution style defaults to "even"}

\item{distkey}{character. optional. Distribution key}

\item{compound_sort}{character vector. optional. Compound sort keys}

\item{interleaved_sort}{character vector. optional. Interleaved sort keys}
}
\value{
character
}
\description{
Get Table Attributes String
}
