% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{temp_table_name}
\alias{temp_table_name}
\title{Generate names for temp tables}
\usage{
temp_table_name(n = 1, prefix = "tt_")
}
\arguments{
\item{n}{integer. The number of names to generate}

\item{prefix}{character. The prefix to use for the table name, defaults to tt_}
}
\description{
Generates uuids (without dashes) to use as names for temp tables and adds a prefix.
}
