% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{update_column_types}
\alias{update_column_types}
\title{Update column types to what we auto-detect}
\usage{
update_column_types(data, dbcon, table_name)
}
\arguments{
\item{data}{Data that we'll use to sense column types (and name columns)}

\item{dbcon}{Database connection to Redshift}

\item{table_name}{Preferably a dbplyr::in_schema specification of table_name}
}
\value{
Not specified, function for side effect
}
\description{
Update column types to what we auto-detect
}
