% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{uploadToS3}
\alias{uploadToS3}
\title{uploadToS3 handles the -test thing on its own since it uses the zapieR methods}
\usage{
uploadToS3(data, bucket, split_files)
}
\arguments{
\item{data}{data.frame}

\item{bucket}{character}

\item{split_files}{Number of files to split the data.frame into when uploading to S3}
}
\description{
uploadToS3 handles the -test thing on its own since it uses the zapieR methods
}
