% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-utils.R
\name{view_definition}
\alias{view_definition}
\title{Show definition for a view}
\usage{
view_definition(dbcon, view_name)
}
\arguments{
\item{dbcon}{A database connection object}

\item{view_name}{The view you'd like to describe}
}
\value{
character
}
\description{
Show definition for a view
}
