% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_utils.R
\name{rs_list_tables}
\alias{rs_list_tables}
\title{List redshift tables}
\usage{
rs_list_tables(dbcon)
}
\arguments{
\item{dbcon}{an RPostgres/RJDBC connection to the redshift server}
}
\value{
a data.frame of table names
}
\description{
Lists all tables available in redshift database
}
