% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_utils.R
\name{rs_table_def}
\alias{rs_table_def}
\title{Show table information}
\usage{
rs_table_def(dbcon, table, schema = "public")
}
\arguments{
\item{dbcon}{an RPostgres/RJDBC connection to the redshift server}

\item{table}{table name}

\item{schema}{schema. defaults to public}
}
\value{
a data.frame containing table information
}
\description{
Show table information
}
